[#2094] Move actor data preparation into data models
Moves skill, tool, armor class, initiative, spellcasting ability,
condition immunity, and scale value preparation into data models.
This leaves only spellcasting preparation in the main actor class.
