+++ module/applications/activity/attack-sheet.mjs
@@ -1,4 +1,3 @@
-import { safePropertyExists } from "../../utils.mjs";
 import ActivitySheet from "./activity-sheet.mjs";
 
 /**
