+++ module/applications/advancement/advancement-selection.mjs
@@ -1,5 +1,3 @@
-import Advancement from "../../documents/advancement/advancement.mjs";
-
 /**
  * Presents a list of advancement types to create when clicking the new advancement button.
  * Once a type is selected, this hands the process over to the advancement's individual editing interface.
