+++ module/data/abstract.mjs
@@ -375,6 +375,33 @@ export class ActorDataModel extends SystemDataModel {
   }
 
   /* -------------------------------------------- */
+  /*  Data Preparation                            */
+  /* -------------------------------------------- */
+
+  /**
+   * Data preparation steps to perform after item data has been prepared, but before active effects are applied.
+   */
+  prepareEmbeddedData() {
+    this._prepareScaleValues();
+  }
+
+  /* -------------------------------------------- */
+
+  /**
+   * Derive any values that have been scaled by the Advancement system.
+   * Mutates the value of the `system.scale` object.
+   * @protected
+   */
+  _prepareScaleValues() {
+    this.scale = this.parent.items.reduce((scale, item) => {
+      if ( CONFIG.DND5E.advancementTypes.ScaleValue.validItemTypes.has(item.type) ) {
+        scale[item.identifier] = item.scaleValues;
+      }
+      return scale;
+    }, {});
+  }
+
+  /* -------------------------------------------- */
   /*  Helpers                                     */
   /* -------------------------------------------- */
