+++ module/data/shared/movement-field.mjs
@@ -1,6 +1,17 @@
 const { BooleanField, NumberField, StringField } = foundry.data.fields;
 
 /**
+ * @typedef {object} MovementData
+ * @property {number} burrow  Actor burrowing speed.
+ * @property {number} climb   Actor climbing speed.
+ * @property {number} fly     Actor flying speed.
+ * @property {number} swim    Actor swimming speed.
+ * @property {number} walk    Actor walking speed.
+ * @property {string} units   Movement used to measure the various speeds.
+ * @property {boolean} hover  This flying creature able to hover in place.
+ */
+
+/**
  * Field for storing movement data.
  */
 export default class MovementField extends foundry.data.fields.SchemaField {
