+++ module/data/shared/senses-field.mjs
@@ -1,6 +1,16 @@
 const { NumberField, StringField } = foundry.data.fields;
 
 /**
+ * @typedef {object} SensesData
+ * @property {number} darkvision       Creature's darkvision range.
+ * @property {number} blindsight       Creature's blindsight range.
+ * @property {number} tremorsense      Creature's tremorsense range.
+ * @property {number} truesight        Creature's truesight range.
+ * @property {string} units            Distance units used to measure senses.
+ * @property {string} special          Description of any special senses or restrictions.
+ */
+
+/**
  * Field for storing senses data.
  */
 export default class SensesField extends foundry.data.fields.SchemaField {
